requirements.txt
google-cloud-aiplatform==1.38.1
opencv-python==4.9.0.80
numpy==1.26.3
python-docx==1.0.0 